DocList をコンポネント化してドキュメントを表示する設定を簡単にする
