+++ src/components/Docs.tsx
@@ -1,73 +1,10 @@
-import { useState, useEffect } from "react";
-import supabase from "../lib/supabase";
-import Pagination from "./Pagination";
+import DocList from "./DocList";
 
-interface Doc {
-  body: string;
-  created_at: string | null;
-  id: number;
-  last_updated_user_id: number;
-  title: string;
-  updated_at: string | null;
-  user_id: number;
-}
-
-const itemsPerPage = 1;
-
-export default function Docs() {
-  const [docs, setDocs] = useState<Doc[]>([]);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(0);
-  const [loading, setLoading] = useState(true);
-
-  const fetchDocs = async (page: number) => {
-    setLoading(true);
-
-    const start = (page - 1) * itemsPerPage;
-    const end = start + itemsPerPage - 1;
-
-    const { data, count, error } = await supabase
-      .from("docs")
-      .select("*", { count: "exact" })
-      .range(start, end);
-
-    if (error) {
-      console.error("Error fetching docs:", error.message);
-      setLoading(false);
-      return;
-    }
-
-    setDocs(data || []);
-    setTotalPages(Math.ceil((count || 0) / itemsPerPage));
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    fetchDocs(currentPage);
-  }, [currentPage]);
-
+export default function DocsPage() {
   return (
     <div style={{ padding: "20px" }}>
       <h1>Docs</h1>
-
-      {loading ? (
-        <p>Loading...</p>
-      ) : (
-        <ul>
-          {docs.map((doc) => (
-            <li key={doc.id}>
-              <h3>{doc.title}</h3>
-              <p>{doc.body}</p>
-            </li>
-          ))}
-        </ul>
-      )}
-
-      <Pagination
-        currentPage={currentPage}
-        totalPages={totalPages}
-        onPageChange={(page) => setCurrentPage(page)}
-      />
+      <DocList itemsPerPage={2} />
     </div>
   );
 }
